+++ src/event/conn.rs
@@ -35,6 +35,7 @@ macro_rules! find_conn_attr_by_kind {
 
 #[derive(Serialize, Deserialize)]
 pub struct PortScanFields {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
     pub dst_ports: Vec<u16>,
@@ -49,8 +50,9 @@ impl PortScanFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             self.dst_addr.to_string(),
             vector_to_string(&self.dst_ports),
@@ -65,6 +67,7 @@ impl PortScanFields {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct PortScan {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
@@ -81,7 +84,8 @@ impl fmt::Display for PortScan {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(
             f,
-            "src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} triage_scores={:?}",
+            "sensor={:?} src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} triage_scores={:?}",
+            self.sensor,
             self.src_addr.to_string(),
             self.dst_addr.to_string(),
             vector_to_string(&self.dst_ports),
@@ -96,6 +100,7 @@ impl fmt::Display for PortScan {
 impl PortScan {
     pub(super) fn new(time: DateTime<Utc>, fields: &PortScanFields) -> Self {
         PortScan {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_addr: fields.dst_addr,
@@ -174,6 +179,7 @@ impl Match for PortScan {
 
 #[derive(Serialize, Deserialize)]
 pub struct MultiHostPortScanFields {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_port: u16,
     pub dst_addrs: Vec<IpAddr>,
@@ -188,8 +194,9 @@ impl MultiHostPortScanFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             vector_to_string(&self.dst_addrs),
             self.dst_port.to_string(),
@@ -204,6 +211,7 @@ impl MultiHostPortScanFields {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct MultiHostPortScan {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_port: u16,
@@ -220,7 +228,8 @@ impl fmt::Display for MultiHostPortScan {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(
             f,
-            "src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} triage_scores={:?}",
+            "sensor={:?} src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} triage_scores={:?}",
+            self.sensor,
             self.src_addr.to_string(),
             vector_to_string(&self.dst_addrs),
             self.dst_port.to_string(),
@@ -235,6 +244,7 @@ impl fmt::Display for MultiHostPortScan {
 impl MultiHostPortScan {
     pub(super) fn new(time: DateTime<Utc>, fields: &MultiHostPortScanFields) -> Self {
         MultiHostPortScan {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_port: fields.dst_port,
@@ -311,6 +321,7 @@ impl Match for MultiHostPortScan {
 
 #[derive(Serialize, Deserialize)]
 pub struct ExternalDdosFields {
+    pub sensor: String,
     pub src_addrs: Vec<IpAddr>,
     pub dst_addr: IpAddr,
     pub proto: u8,
@@ -324,8 +335,9 @@ impl ExternalDdosFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             vector_to_string(&self.src_addrs),
             self.dst_addr.to_string(),
             self.proto.to_string(),
@@ -339,6 +351,7 @@ impl ExternalDdosFields {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct ExternalDdos {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addrs: Vec<IpAddr>,
     pub dst_addr: IpAddr,
@@ -354,7 +367,8 @@ impl fmt::Display for ExternalDdos {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(
             f,
-            "src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} triage_scores={:?}",
+            "sensor={:?} src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} triage_scores={:?}",
+            self.sensor,
             vector_to_string(&self.src_addrs),
             self.dst_addr.to_string(),
             self.proto.to_string(),
@@ -368,6 +382,7 @@ impl fmt::Display for ExternalDdos {
 impl ExternalDdos {
     pub(super) fn new(time: DateTime<Utc>, fields: &ExternalDdosFields) -> Self {
         ExternalDdos {
+            sensor: fields.sensor.clone(),
             time,
             src_addrs: fields.src_addrs.clone(),
             dst_addr: fields.dst_addr,